+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,6 +1,5 @@
 package com.techcourse.dao;
 
-import java.sql.Connection;
 import java.sql.ResultSet;
 import java.util.List;
 
@@ -34,9 +33,9 @@ public class UserDao {
         jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail());
     }
 
-    public void update(final Connection connection, final User user) {
+    public void update(final User user) {
         final var sql = "update users set password = ? where id = ?";
-        jdbcTemplate.update(connection, sql, user.getPassword(), user.getId());
+        jdbcTemplate.update(sql, user.getPassword(), user.getId());
     }
 
     public List<User> findAll() {
