+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -1,7 +1,5 @@
 package com.techcourse.dao;
 
-import java.sql.Connection;
-
 import javax.sql.DataSource;
 
 import com.interface21.jdbc.core.JdbcTemplate;
@@ -19,10 +17,10 @@ public class UserHistoryDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public void log(final Connection connection, final UserHistory userHistory) {
+    public void log(final UserHistory userHistory) {
         final var sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
 
-        jdbcTemplate.update(connection, sql,
+        jdbcTemplate.update(sql,
                 userHistory.getUserId(), userHistory.getAccount(), userHistory.getPassword(),
                 userHistory.getEmail(), userHistory.getCreatedAt(), userHistory.getCreateBy());
     }
