+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,47 +1,12 @@
 package com.techcourse.service;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.interface21.jdbc.core.TransactionManager;
-import com.interface21.jdbc.exception.TransactionExecutionException;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionManager transactionManager;
+    User findById(long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao,
-                       final TransactionManager transactionManager) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionManager = transactionManager;
-    }
+    void insert(User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        try {
-            transactionManager.execute(connection -> {
-                final var user = findById(id);
-                user.changePassword(newPassword);
-                userDao.update(connection, user);
-                userHistoryDao.log(connection, new UserHistory(user, createBy));
-            });
-        } catch (final TransactionExecutionException e) {
-            log.info("비밀번호 변경에 실패하였습니다.");
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
